live-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
paho-mqtt>=1.6.1
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
